Refactoring, renaming formatter helper class to more generic name
